Moved Sampler package to `com.linecorp.armeria.common.util`
