+++ core/src/main/java/com/linecorp/armeria/client/logging/LoggingClient.java
@@ -33,7 +33,7 @@ import com.linecorp.armeria.common.Request;
 import com.linecorp.armeria.common.Response;
 import com.linecorp.armeria.common.logging.LogLevel;
 import com.linecorp.armeria.common.logging.RequestLogAvailability;
-import com.linecorp.armeria.internal.logging.Sampler;
+import com.linecorp.armeria.common.util.Sampler;
 import com.linecorp.armeria.server.logging.LoggingService;
 
 /**
