+++ core/src/main/java/com/linecorp/armeria/client/logging/LoggingClientBuilder.java
@@ -22,7 +22,7 @@ import com.linecorp.armeria.client.Client;
 import com.linecorp.armeria.common.Request;
 import com.linecorp.armeria.common.Response;
 import com.linecorp.armeria.common.logging.LoggingDecoratorBuilder;
-import com.linecorp.armeria.internal.logging.Sampler;
+import com.linecorp.armeria.common.util.Sampler;
 
 /**
  * Builds a new {@link LoggingClient}.
